Renaming course-page-resources to pages-and-resources
It’s all course stuff in the end.
