+++ src/CourseAuthoringRoutes.jsx
@@ -4,7 +4,7 @@ import { Switch, useRouteMatch } from 'react-router';
 import { PageRoute } from '@edx/frontend-platform/react';
 
 import CourseAuthoringPage from './CourseAuthoringPage';
-import { CoursePageResources } from './course-page-resources';
+import { CoursePageResources } from './pages-and-resources';
 import ProctoredExamSettings from './proctored-exam-settings/ProctoredExamSettings';
 import DiscussionToolSelectorContainer from './discussion-tool-selector/DiscussionToolSelectorContainer';
 
@@ -28,10 +28,10 @@ export default function CourseAuthoringRoutes({ courseId }) {
   return (
     <CourseAuthoringPage courseId={courseId}>
       <Switch>
-        <PageRoute exact path={`${path}/pages`}>
+        <PageRoute exact path={`${path}/pages-and-resources`}>
           <CoursePageResources courseId={courseId} />
         </PageRoute>
-        <PageRoute path={`${path}/pages/discussion`}>
+        <PageRoute path={`${path}/pages-and-resources/discussion`}>
           <DiscussionToolSelectorContainer courseId={courseId} />
         </PageRoute>
         <PageRoute path={`${path}/proctored-exam-settings`}>
