+++ src/index.scss
@@ -1,5 +1,5 @@
 
-@import './course-page-resources/index.scss';
+@import './pages-and-resources/index.scss';
 @import "~@edx/brand/paragon/fonts";
 @import "~@edx/brand/paragon/variables";
 @import "~@edx/paragon/scss/core/core";
