+++ src/pages-and-resources/course-page/CoursePageConfigCard.jsx
