+++ src/pages-and-resources/course-page/CoursePageConfigCard.test.jsx
