+++ src/pages-and-resources/index.scss
